backend/requirements.txt
Flask==3.1.2
Flask-Cors==4.0.1
PyJWT==2.9.0
Werkzeug==3.1.3
